Fixed the game caption (title for window).
Did some reorganizing with game states and how they work. Refining the game
logic a bit, as I was a bit unsure of where to put everything. Seems best to
use the Game object for general Game stuff (OBVIOUSLY, RIGHT?).

May have went overboard with this release. I added a pause screen, I added a
fade which is questionably implemented, though neat. I "fixed" the audio
source memory leak. By which I mean I was a moron last commit. I still have
to figure out a way to handle this sound playing multiple times at once,
because it seems to not be playable if it's already playing. I've opted to
simply "stop" the sound before attempting to replay it. Seems to work
alright.

Added a song that plays in the background, turned off thuds for the
IntroGameBoard, as it was annoying as hell and only for display purposes
anyway.

Changed the main menu overlay graphic, and made the game pieces rounded a
bit. Though this may change.