Tetrominoes are in.
The game now functions like a generic Tetris game. I was having a lot of
trouble getting the math to produce consistent results with X and Y values
allowing for decimal values (when shifting, I seemed to lose some level of
precision, so some of the pieces were shifting over a pixel when they
weren't supposed to). For the time being, it's just your average Tetris.

Second problem I need to fix is exporting the math used in shifting along a
pivot, so I can check if any of the pieces collide or go outside the edge of
the map. That'll be the next step, I suppose. But it's 12AM, so I'll put
that off for tomorrow.