Changed the screen resolution to 640x640. It needed to be taller.
Added automatic sliding (when you hold left and right, it slides at a
reasonable pace).

Added the square (O) and long (I) tetrominoes.

Rewrote the pivot logic to be more accessible to outside functions, so I can
know which piece will be where when it pivots. This is mostly to check for
collisions before doing an actual pivot.

Made it so pieces and slide just before snapping to the bottom of the
screen.