Moved credits to separate file (as opposed to just being in the source as comments.
Added extension to table functions -- specifically a version if pairs() and
ipairs() that doesn't iterate through the source table, but instead iterates
through a copy of it. In case you didn't know, the standard iterator is not
exactly "safe." If you modify the source table while iterating, it will
directly affect the iterator. As such, I have called these functions
table.safeIPairs() and table.safePairs(). I also added a generic copier
called table.copy().

The game now has a semblence of a score, and standard tetris play.

All I have to do is add a game-over screen, and this is basically a game.